project/somaz-bigquery/cloud-functions/somaz-cdn-bucket-file-download/requirements.txt
google-cloud-storage
google-cloud-logging
